requirements.txt
certifi==2019.3.9
chardet==3.0.4
cognitive-face==1.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
djangorestframework==3.9.3
gunicorn==19.9.0
idna==2.5
newsapi-python==0.2.3
numpy==1.16.3
oauthlib==3.0.1
opencv-contrib-python==4.1.0.25
opencv-python==3.3.0.9
Pillow==6.0.0
psycopg2==2.8.2
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2019.1
requests==2.17.1
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.21.1
whitenoise==4.1.2
